UI now split in two files. ui.c re-changed from base 0.5.0 from the original OP repo. should make merge to 0.5.1 a LOT smoother as well as future upgrades
