+++ selfdrive/car/ford/.gitignore
@@ -1,0 +1,2 @@
+buttons.msg
+buttons.cc.msg
